Location Manager Project: implementation of the tab view and base of permanent request screen
