+++ Projects/Location Manager/Location Manager/FixedPermanentRequestView.swift
@@ -1,0 +1,27 @@
+//
+//  File name: FixedPermanentRequestView.swift
+//  Project name: Location Manager
+//  Workspace name: SwiftUI-Samples
+//
+//  Created by: nothing-to-add on 18/09/2023
+//  Using Swift 5.0
+//  Copyright (c) 2023 nothing-to-add
+//
+
+import SwiftUI
+
+struct FixedPermanentRequestView: View {
+    
+    var body: some View {
+        VStack {
+            Text("Fixed Permanent Request View")
+        }
+    }
+}
+
+struct FixedPermanentRequestView_Previews: PreviewProvider {
+    static var previews: some View {
+        FixedPermanentRequestView()
+    }
+}
+
